Added consistent hashing within a single process
This introduces a  feature "consistent". The feature  will replace the
default random state  hasher with a default hasher with  a shared seed
for a single process.

This enables the hashes to be deterministic for the whole process. And
this then will  enable composition of hash maps or  sets as set values
or as keys of hash maps. With the drawback of having less robustness
for hash DOS attacks.

So if you  do not use hash sets  or hash maps as keys in  hash maps or
values in sets, then do not enable the feature.